Update schema
Fix type checking
